+++ common4j/src/main/com/microsoft/identity/common/java/nativeauth/providers/responses/jit/JITIntrospectApiResponse.kt
@@ -26,88 +26,95 @@ import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 import com.microsoft.identity.common.java.nativeauth.providers.INativeAuthApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.ApiErrorResult
-import com.microsoft.identity.common.java.nativeauth.util.isInvalidGrant
-import com.microsoft.identity.common.java.nativeauth.util.isOOBValueInvalid
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.AuthenticationMethodApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.toListOfAuthenticationMethodApiResult
 import com.microsoft.identity.common.java.nativeauth.util.isRedirect
+import java.lang.IllegalStateException
 import java.net.HttpURLConnection
 
 /**
- * Represents the raw response from the register/continue endpoint.
- * Can be converted to JITContinueAPIResult using the provided toResult() method.
+ * Represents the raw response from the register/introspect endpoint.
+ * Can be converted to JITIntrospectApiResult using the provided toResult() method.
  */
-class JITContinueNativeAuthApiResponse(
+class JITIntrospectApiResponse(
     @Expose override var statusCode: Int,
     correlationId: String,
     override val continuationToken: String?,
+    @Expose @SerializedName("methods") val methods: List<AuthenticationMethodApiResponse>?,
     override val error: String?,
     override val errorDescription: String?,
-    @SerializedName("suberror") val subError: String?,
+    @SerializedName("error_uri") val errorUri: String?,
     @SerializedName("error_codes") val errorCodes: List<Int>?,
     override val challengeType: String?,
     override val redirectReason: String?,
 ) : INativeAuthApiResponse(statusCode, correlationId, continuationToken, challengeType, redirectReason, error, errorDescription) {
+
     override fun toUnsanitizedString(): String {
-        return "JITContinueAPIResponse(statusCode=$statusCode, " +
+        return "JITIntrospectApiResponse(statusCode=$statusCode, " +
                 "correlationId=$correlationId " +
-                "error=$error, errorCodes=$errorCodes, errorDescription=$errorDescription," +
+                "error=$error, errorCodes=$errorCodes, errorDescription=$errorDescription, " +
+                "methods=$methods, challengeType=$challengeType, " +
                 "redirectReason=$redirectReason)"
     }
 
-    override fun toString(): String = "JITContinueAPIResponse(statusCode=$statusCode, " +
-            "correlationId=$correlationId)"
+    override fun toString(): String = "JITIntrospectApiResponse(statusCode=$statusCode, " +
+            "correlationId=$correlationId"
 
-    fun toResult(): JITContinueApiResult {
+    fun toResult(): JITIntrospectApiResult {
         return when (statusCode) {
             // Handle 400 errors
             HttpURLConnection.HTTP_BAD_REQUEST -> {
-                return when {
-                    error.isInvalidGrant() && subError.isOOBValueInvalid() -> {
-                        JITContinueApiResult.CodeIncorrect(
-                            correlationId = correlationId,
-                            error = error.orEmpty(),
-                            errorDescription = errorDescription.orEmpty(),
-                            errorCodes = errorCodes.orEmpty(),
-                            subError = subError.orEmpty()
-                        )
-                    }
-                    else -> {
-                        JITContinueApiResult.UnknownError(
-                            error = error.orEmpty(),
-                            errorDescription = errorDescription.orEmpty(),
-                            errorCodes = errorCodes.orEmpty(),
-                            correlationId = correlationId
-                        )
-                    }
-                }
+                JITIntrospectApiResult.UnknownError(
+                    error = error.orEmpty(),
+                    errorDescription = errorDescription.orEmpty(),
+                    errorCodes = errorCodes.orEmpty(),
+                    correlationId = correlationId
+                )
             }
 
             // Handle success and redirect
             HttpURLConnection.HTTP_OK -> {
                 return when {
                     challengeType.isRedirect() -> {
-                        JITContinueApiResult.Redirect(
+                        JITIntrospectApiResult.Redirect(
                             correlationId = correlationId,
                             redirectReason = redirectReason.orEmpty()
                         )
                     }
-                    continuationToken.isNullOrBlank() -> {
-                        JITContinueApiResult.UnknownError(
+                    methods.isNullOrEmpty() -> {
+                        JITIntrospectApiResult.UnknownError(
                             error = ApiErrorResult.INVALID_STATE,
-                            errorDescription = "Register authentication method /continue did not return continuationToken field",
+                            errorDescription = "register/introspect did not return methods",
                             errorCodes = errorCodes.orEmpty(),
                             correlationId = correlationId
                         )
                     }
                     else -> {
-                        JITContinueApiResult.Success(
-                            correlationId = correlationId,
-                            continuationToken = continuationToken
-                        )
+                        try {
+                            JITIntrospectApiResult.Success(
+                                correlationId = correlationId,
+                                continuationToken = continuationToken
+                                    ?: return JITIntrospectApiResult.UnknownError(
+                                        error = ApiErrorResult.INVALID_STATE,
+                                        errorDescription = "register/introspect did not return a continuation token",
+                                        errorCodes = errorCodes.orEmpty(),
+                                        correlationId = correlationId
+                                    ),
+                                methods = methods.toListOfAuthenticationMethodApiResult()
+                            )
+                        } catch (e: IllegalStateException) {
+                            JITIntrospectApiResult.UnknownError(
+                                error = ApiErrorResult.INVALID_STATE,
+                                errorDescription = "register/introspect did not return valid methods: ${e.message}",
+                                errorCodes = errorCodes.orEmpty(),
+                                correlationId = correlationId
+                            )
+                        }
                     }
                 }
             }
             else -> {
-                JITContinueApiResult.UnknownError(
+                JITIntrospectApiResult.UnknownError(
                     error = error.orEmpty(),
                     errorDescription = errorDescription.orEmpty(),
                     errorCodes = errorCodes.orEmpty(),
