+++ common4j/src/test/com/microsoft/identity/common/java/nativeauth/providers/NativeAuthResponseHandlerTest.kt
@@ -27,36 +27,36 @@ import com.microsoft.identity.common.java.nativeauth.providers.NativeAuthRespons
 import com.microsoft.identity.common.java.nativeauth.providers.responses.ApiErrorResult
 import com.microsoft.identity.common.java.nativeauth.providers.responses.UserAttributeApiResult
 import com.microsoft.identity.common.java.nativeauth.providers.responses.UserAttributeOptionsApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITChallengeNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITChallengeApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITChallengeApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITContinueNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITContinueApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITContinueApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITIntrospectNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITIntrospectApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.jit.JITIntrospectApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordChallengeNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordChallengeApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordChallengeApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordContinueNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordContinueApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordContinueApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordPollCompletionNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordPollCompletionApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordPollCompletionApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordStartNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordStartApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordStartApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordSubmitNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordSubmitApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.resetpassword.ResetPasswordSubmitApiResult
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.AuthenticationMethodApiResponse
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInChallengeNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInChallengeApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInChallengeApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInInitiateNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInInitiateApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInInitiateApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInIntrospectNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInIntrospectApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInIntrospectApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInTokenNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInTokenApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signin.SignInTokenApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpChallengeNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpChallengeApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpChallengeApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpContinueNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpContinueApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpContinueApiResult
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpStartNativeAuthApiResponse
+import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpStartApiResponse
 import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.SignUpStartApiResult
 import com.microsoft.identity.common.java.net.HttpResponse
 import com.microsoft.identity.common.nativeauth.ApiConstants
@@ -191,7 +191,7 @@ class NativeAuthResponseHandlerTest {
     // region SignUp Start
     @Test
     fun testSignUpStartApiResponseVerificationRequired() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -215,7 +215,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseUnknownErrorWithUnauthorizedClientError() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -235,7 +235,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseVerificationRequiredWithMissingSignUpToken() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -255,7 +255,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseVerificationRequiredWithMissingUnverifiedAttributes() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -275,7 +275,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseUserAlreadyExists() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -294,7 +294,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseInvalidEmail() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -313,7 +313,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseInvalidParameter() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -332,7 +332,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseUnsupportedChallengeType() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -351,7 +351,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponsePasswordTooWeak() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -370,7 +370,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponsePasswordTooLong() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -389,7 +389,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponsePasswordTooShort() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -408,7 +408,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponsePasswordBanned() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -427,7 +427,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponsePasswordRecentlyUsed() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -446,7 +446,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseRedirect() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -467,7 +467,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseAuthNotSupported() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -486,7 +486,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseValidationFailed() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -505,7 +505,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpStartApiResponseUncommonErrorStatusCode() {
-        val signUpStartApiResponse = SignUpStartNativeAuthApiResponse(
+        val signUpStartApiResponse = SignUpStartApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -527,7 +527,7 @@ class NativeAuthResponseHandlerTest {
     // region SignUp Challenge
     @Test
     fun testSignUpChallengeApiResponseUnsupportedChallengeType() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -547,7 +547,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseExpiredToken() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -567,7 +567,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseOOBRequired() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -591,7 +591,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseOOBRequiredWithCodeLengthMissing() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -612,7 +612,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseOOBRequiredWithChallengeTargetLabelMissing() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -632,7 +632,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseOOBRequiredWithChallengeChannelMissing() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -652,7 +652,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponsePasswordRequired() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = passwordChallengeType,
             redirectReason = null,
@@ -676,7 +676,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseRedirect() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -697,7 +697,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseIncorrectChallengeType() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = invalidChallengeType,
             redirectReason = null,
@@ -717,7 +717,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseWithNoSContinuationToken() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -737,7 +737,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponsePasswordWithNoContinuationToken() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = passwordChallengeType,
             redirectReason = null,
@@ -757,7 +757,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpChallengeApiResponseUncommonStatusCode() {
-        val signUpChallengeApiResponse = SignUpChallengeNativeAuthApiResponse(
+        val signUpChallengeApiResponse = SignUpChallengeApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -779,7 +779,7 @@ class NativeAuthResponseHandlerTest {
     // region SignUp Continue
     @Test
     fun testSignUpContinueApiResponse() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             error = null,
@@ -801,7 +801,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiResponseRedirect() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             error = null,
@@ -823,7 +823,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiResponseNoSLT() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             error = null,
@@ -845,7 +845,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiResponseErrorWithNoContinuationToken() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = null,
@@ -866,7 +866,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiResponseWithUnknownError() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = nullString,
@@ -887,7 +887,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiPasswordTooWeak() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -908,7 +908,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiPasswordTooLong() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -929,7 +929,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiPasswordTooShort() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -950,7 +950,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiPasswordBanned() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -971,7 +971,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiPasswordRecentlyUsed() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -992,7 +992,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUserAlreadyExists() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = userAlreadyExistsError,
@@ -1013,7 +1013,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiInvalidOOBWithErrorCode1() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -1034,7 +1034,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiInvalidOOBWithErrorCode2() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -1055,7 +1055,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiInvalidOOBWithErrorCode3() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -1076,7 +1076,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiInvalidRequestNullErrorCode() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidRequestError,
@@ -1097,7 +1097,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiInvalidRequestInvalidErrorCode() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidRequestError,
@@ -1118,7 +1118,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiExpiredToken() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = expiredTokenError,
@@ -1139,7 +1139,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiAttributeValidationFailed() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -1160,7 +1160,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiAttributeValidationFailedMissingInvalidAttributes() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -1181,7 +1181,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUserAttributesRequiredResponse() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = attributesRequiredError,
@@ -1203,7 +1203,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUserInvalidAttributesResponse() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = invalidGrantError,
@@ -1225,7 +1225,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUserAttributesRequiredResponseNoFlowToken() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = attributeValidationFailed,
@@ -1246,7 +1246,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUserAttributesRequiredResponseNoRequiredAttributes() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = attributesRequiredError,
@@ -1267,7 +1267,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiAuthenticationRequiredResponse() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = credentialRequiredError,
@@ -1289,7 +1289,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignUpContinueApiUncommonStatusCode() {
-        val signUpContinueApiResponse = SignUpContinueNativeAuthApiResponse(
+        val signUpContinueApiResponse = SignUpContinueApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             error = null,
@@ -1312,7 +1312,7 @@ class NativeAuthResponseHandlerTest {
     // region ResetPassword Start
     @Test
     fun testValidateSsprStartResultWithSuccessReturnChallengeType() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             challengeType = redirect,
@@ -1330,7 +1330,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultWithSuccessReturnContinuationToken() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = null,
@@ -1348,7 +1348,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSsprStartResultUnknownErrorWithUnauthorizedClientError() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1366,7 +1366,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultWithSuccessNoRedirectButMissingToken() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1383,7 +1383,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultUserNotFound() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1400,7 +1400,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultInvalidGrant() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1417,7 +1417,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultUnsupportedChallengeType() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1434,7 +1434,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultUnknownError() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1452,7 +1452,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprStartResultUncommonStatusCode() {
-        val resetPasswordStartApiResponse = ResetPasswordStartNativeAuthApiResponse(
+        val resetPasswordStartApiResponse = ResetPasswordStartApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1472,7 +1472,7 @@ class NativeAuthResponseHandlerTest {
     // region ResetPassword Challenge
     @Test
     fun testValidateSsprChallengeResultSuccessWithOobChallenge() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = oobChallengeType,
@@ -1497,7 +1497,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessWithPasswordChallenge() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = 200,
             continuationToken = "1234",
             challengeType = passwordChallengeType,
@@ -1519,7 +1519,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessWithRedirectChallenge() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = redirect,
@@ -1542,7 +1542,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessAndMissingContinuationToken() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             challengeType = oobChallengeType,
@@ -1564,7 +1564,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessAndMissingChallengeTargetLabel() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = oobChallengeType,
@@ -1586,7 +1586,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessAndMissingChallengeChannel() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = oobChallengeType,
@@ -1608,7 +1608,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessAndMissingChallengeType() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = null,
@@ -1630,7 +1630,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultSuccessAndMissingCodeLength() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = oobChallengeType,
@@ -1652,7 +1652,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultExpiredToken() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1674,7 +1674,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultUnsupportedChallengeType() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1696,7 +1696,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultInvalidGrant() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1718,7 +1718,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprChallengeResultUncommonStatusCode() {
-        val resetPasswordChallengeApiResponse = ResetPasswordChallengeNativeAuthApiResponse(
+        val resetPasswordChallengeApiResponse = ResetPasswordChallengeApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1742,7 +1742,7 @@ class NativeAuthResponseHandlerTest {
     // region ResetPassword Continue
     @Test
     fun testValidateSsprContinueResultWithSuccessWithSubmitToken() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = null,
@@ -1763,7 +1763,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultSuccessNullExpiresIn() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             challengeType = null,
@@ -1783,7 +1783,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultSuccessNullSubmitToken() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1802,7 +1802,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultSuccessRedirect() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             challengeType = redirect,
@@ -1822,7 +1822,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultOOBIncorrect() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1841,7 +1841,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultNoErrorCodes() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1860,7 +1860,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultExpiredToken() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1879,7 +1879,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultNoErrorName() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1898,7 +1898,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprContinueResultUncommonStatusCode() {
-        val resetPasswordContinueApiResponse = ResetPasswordContinueNativeAuthApiResponse(
+        val resetPasswordContinueApiResponse = ResetPasswordContinueApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             challengeType = null,
@@ -1919,7 +1919,7 @@ class NativeAuthResponseHandlerTest {
     // region ResetPassword Submit
     @Test
     fun testValidateSsprSubmitResultSuccessStartPolling() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             pollInterval = pollInterval,
@@ -1940,7 +1940,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultSuccessRedirect() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             pollInterval = pollInterval,
@@ -1960,7 +1960,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultWithSuccessAndMissingContinuationToken() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             pollInterval = pollInterval,
@@ -1979,7 +1979,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultWithSuccessAndMissingPollInterval() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             pollInterval = null,
@@ -1999,7 +1999,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultExpiredToken() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2018,7 +2018,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultPasswordTooWeak() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2037,7 +2037,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultPasswordTooLong() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2056,7 +2056,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultPasswordTooShort() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2075,7 +2075,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultPasswordBanned() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2094,7 +2094,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultPasswordRecentlyUsed() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2113,7 +2113,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultUnknownError() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2134,7 +2134,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprSubmitResultUncommonStatusCode() {
-        val resetPasswordSubmitApiResponse = ResetPasswordSubmitNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordSubmitApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             pollInterval = null,
@@ -2157,7 +2157,7 @@ class NativeAuthResponseHandlerTest {
     // region ResetPassword PollCompletion
     @Test
     fun testValidateSsprPollCompletionResultSucceeded() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = successStatusCode,
             status = succeededStatus,
             continuationToken = null,
@@ -2177,7 +2177,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultInProgress() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = successStatusCode,
             status = inProgressStatus,
             continuationToken = null,
@@ -2197,7 +2197,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPollingFailed() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = successStatusCode,
             status = failedStatus,
             continuationToken = null,
@@ -2217,7 +2217,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPollingSuccessRedirect() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = successStatusCode,
             status = failedStatus,
             continuationToken = null,
@@ -2238,7 +2238,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPasswordBanned() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = invalidGrantError,
             errorDescription = null,
@@ -2258,7 +2258,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPasswordTooShort() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = invalidGrantError,
             errorDescription = null,
@@ -2278,7 +2278,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPasswordTooLong() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = invalidGrantError,
             errorDescription = null,
@@ -2298,7 +2298,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPasswordRecentlyUsed() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = invalidGrantError,
             errorDescription = null,
@@ -2318,7 +2318,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultPasswordTooWeak() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = invalidGrantError,
             errorDescription = null,
@@ -2338,7 +2338,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionExpiredToken() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = expiredTokenError,
             errorDescription = null,
@@ -2358,7 +2358,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionExplicitUserNotFound() {
-        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordSubmitApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = errorStatusCode,
             error = userNotFoundError,
             errorDescription = null,
@@ -2378,7 +2378,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultWithSuccessAndMissingStatus() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = successStatusCode,
             status = null,
             continuationToken = null,
@@ -2398,7 +2398,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testValidateSsprPollCompletionResultUncommonStatusCode() {
-        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionNativeAuthApiResponse(
+        val resetPasswordPollCompletionApiResponse = ResetPasswordPollCompletionApiResponse(
             statusCode = uncommonErrorStatusCode,
             status = null,
             continuationToken = null,
@@ -2420,7 +2420,7 @@ class NativeAuthResponseHandlerTest {
     // region SignIn Initiate
     @Test
     fun testSignInInitiateResultWithRedirectChallenge() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -2439,7 +2439,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseWithSuccess() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2458,7 +2458,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseWithMissingContinuationToken() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2476,7 +2476,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseUnknownErrorWithUnauthorizedClientError() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2494,7 +2494,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseInvalidGrantWithMissingErrorCodes() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2512,7 +2512,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseUserNotFound() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2532,7 +2532,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseUnknownErrorCode() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2550,7 +2550,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateApiResponseWithUnknownError() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2570,7 +2570,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInInitiateUncommonStatusCode() {
-        val signInInitiateApiResponse = SignInInitiateNativeAuthApiResponse(
+        val signInInitiateApiResponse = SignInInitiateApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2590,7 +2590,7 @@ class NativeAuthResponseHandlerTest {
     // region sign in introspect
     @Test
     fun testSignInIntrospectRedirect() {
-        val apiResponse = SignInIntrospectNativeAuthApiResponse(
+        val apiResponse = SignInIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -2623,7 +2623,7 @@ class NativeAuthResponseHandlerTest {
                 challengeChannel = emailChallengeChannel
             )
         )
-        val apiResponse = SignInIntrospectNativeAuthApiResponse(
+        val apiResponse = SignInIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2655,7 +2655,7 @@ class NativeAuthResponseHandlerTest {
                 challengeChannel = emailChallengeChannel
             )
         )
-        val apiResponse = SignInIntrospectNativeAuthApiResponse(
+        val apiResponse = SignInIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2676,7 +2676,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInIntrospectWithEmptyMethodsShouldReturnError() {
-        val apiResponse = SignInIntrospectNativeAuthApiResponse(
+        val apiResponse = SignInIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2709,7 +2709,7 @@ class NativeAuthResponseHandlerTest {
                 challengeChannel = emailChallengeChannel
             )
         )
-        val apiResponse = SignInIntrospectNativeAuthApiResponse(
+        val apiResponse = SignInIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2730,7 +2730,7 @@ class NativeAuthResponseHandlerTest {
     // region SignIn Challenge
     @Test
     fun testSignInChallengeApiResponseInvalidGrant() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2756,7 +2756,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseIntrospectRequired() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2781,7 +2781,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypeOobSuccess() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -2807,7 +2807,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypeOobMissingCodeLength() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -2831,7 +2831,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypeOobMissingChallengeChannel() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -2855,7 +2855,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypeOobMissingChallengeTargetLabel() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -2879,7 +2879,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypePasswordWithMissingContinuationToken() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = passwordChallengeType,
             redirectReason = null,
@@ -2903,7 +2903,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypePassword() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = passwordChallengeType,
             redirectReason = null,
@@ -2928,7 +2928,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseChallengeTypePasswordMissingFlowToken() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = passwordChallengeType,
             redirectReason = null,
@@ -2952,7 +2952,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeApiResponseWithUnknownError() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -2978,7 +2978,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInChallengeUncommonStatusCode() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -3001,7 +3001,7 @@ class NativeAuthResponseHandlerTest {
     }
 
     fun testSignInChallengeSuccessRedirect() {
-        val response = SignInChallengeNativeAuthApiResponse(
+        val response = SignInChallengeApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -3028,7 +3028,7 @@ class NativeAuthResponseHandlerTest {
     // region SignIn Token
     @Test
     fun testSignInTokenApiResponseInvalidGrantMissingErrorCodes() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3047,7 +3047,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseUnknownErrorWithUnauthorizedClientError() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = unauthorizedClientError,
@@ -3064,7 +3064,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseUnknownErrorWithInvalidClientError() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidClientError,
@@ -3081,7 +3081,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseUserNotFound() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = userNotFoundError,
@@ -3100,7 +3100,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseInvalidCredentials() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3119,7 +3119,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseOtpCodeIncorrectWithInvalidGrantErrorCode1() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3139,7 +3139,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseOtpCodeIncorrectWithInvalidGrantErrorCode2() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3159,7 +3159,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseOtpCodeIncorrectWithInvalidGrantErrorCode3() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3179,7 +3179,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseInvalidAuthenticationType() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3198,7 +3198,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseMultipleErrorCodes() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3245,7 +3245,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseWithUnknownError() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = unknownError,
@@ -3264,7 +3264,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseCredentialRequiredExplicitError() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             error = credentialRequiredError,
@@ -3281,7 +3281,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseMFARequired() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = invalidGrantError,
@@ -3302,7 +3302,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseResetPasswordRequired() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = invalidRequestError,
@@ -3323,7 +3323,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testSignInTokenApiResponseUnknownError() {
-        val signInTokenApiResponse = SignInTokenNativeAuthApiResponse(
+        val signInTokenApiResponse = SignInTokenApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = null,
             error = null,
@@ -3700,7 +3700,7 @@ class NativeAuthResponseHandlerTest {
     @Test
     fun testJITIntrospectApiSuccessButEmptyMethodList() {
         val methods = emptyList<AuthenticationMethodApiResponse>()
-        val jitIntrospectApiResponse = JITIntrospectNativeAuthApiResponse(
+        val jitIntrospectApiResponse = JITIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = emailChallengeChannel,
             redirectReason = null,
@@ -3729,7 +3729,7 @@ class NativeAuthResponseHandlerTest {
             challengeChannel = ""
         )
         val methods = listOf(method)
-        val jitIntrospectApiResponse = JITIntrospectNativeAuthApiResponse(
+        val jitIntrospectApiResponse = JITIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = emailChallengeChannel,
             redirectReason = null,
@@ -3758,7 +3758,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITIntrospectApiSuccessRedirect() {
-        val jitIntrospectApiResponse = JITIntrospectNativeAuthApiResponse(
+        val jitIntrospectApiResponse = JITIntrospectApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -3778,7 +3778,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITIntrospectApiResponseBadRequest() {
-        val jitIntrospectApiResponse = JITIntrospectNativeAuthApiResponse(
+        val jitIntrospectApiResponse = JITIntrospectApiResponse(
             statusCode = errorStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -3804,7 +3804,7 @@ class NativeAuthResponseHandlerTest {
     //region JIT challenge
     @Test
     fun testJITChallengeApiSuccessButMissingMandatoryField() {
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = emailChallengeChannel,
             redirectReason = null,
@@ -3831,7 +3831,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITChallengeApiSuccessRedirect() {
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -3855,7 +3855,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITChallengeApiUncommonStatusCodeShouldReturnError() {
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = uncommonErrorStatusCode,
             challengeType = redirect,
             redirectReason = redirectReason,
@@ -3883,7 +3883,7 @@ class NativeAuthResponseHandlerTest {
     @Test
     fun testJITChallengeApiInvalidChallengeTargetIsHandledCorrectly() {
         val errorCodes = listOf(901001)
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -3913,7 +3913,7 @@ class NativeAuthResponseHandlerTest {
     @Test
     fun testJITChallengeApiInvalidRequestWithUnknownErrorCodeIsHandledCorrectly() {
         val errorCodes = listOf(100100)
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = errorStatusCode,
             challengeType = null,
             redirectReason = null,
@@ -3942,7 +3942,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITChallengeApiSuccessIsHandledCorrectly() {
-        val jitChallengeApiResponse = JITChallengeNativeAuthApiResponse(
+        val jitChallengeApiResponse = JITChallengeApiResponse(
             statusCode = successStatusCode,
             challengeType = oobChallengeType,
             redirectReason = null,
@@ -3975,7 +3975,7 @@ class NativeAuthResponseHandlerTest {
     //region JIT continue
     @Test
     fun testJITContinueApiInvalidOOBIsHandledCorrectly() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = invalidGrantError,
@@ -3998,7 +3998,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITContinueApiInvalidGrantGenericSubErrorIsHandledAsUnknownError() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = errorStatusCode,
             continuationToken = continuationToken,
             error = invalidGrantError,
@@ -4021,7 +4021,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITContinueApiUncommonErrorCodeIsHandledAsUnknownError() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = uncommonErrorStatusCode,
             continuationToken = continuationToken,
             error = invalidGrantError,
@@ -4044,7 +4044,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITContinueApiSuccessButMissingContinuationTokenIsHandledAsUnknownError() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = null,
             error = invalidGrantError,
@@ -4066,7 +4066,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITContinueApiSuccessIsHandledAsSuccess() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             error = null,
@@ -4088,7 +4088,7 @@ class NativeAuthResponseHandlerTest {
 
     @Test
     fun testJITContinueApiSuccessRedirect() {
-        val jitContinueApiResponse = JITContinueNativeAuthApiResponse(
+        val jitContinueApiResponse = JITContinueApiResponse(
             statusCode = successStatusCode,
             continuationToken = continuationToken,
             error = null,
